+++ @xen-orchestra/lite/src/pages/host/[uuid]/console.vue
@@ -1,7 +1,7 @@
 <template>
   <div :class="{ 'no-ui': !uiStore.hasUi }" class="host-console-view">
     <div v-if="hasError">{{ t('error-occurred') }}</div>
-    <UiSpinner v-else-if="!isReady" class="spinner" />
+    <UiLoader v-else-if="!isReady" class="loader" />
     <VtsStateHero v-else-if="!isHostRunning" format="page" type="offline" size="large" class="state-hero">
       <span>{{ t('console-offline') }}</span>
       <span class="title typo-h1">{{ t('host-not-running') }}</span>
@@ -24,7 +24,6 @@
 </template>
 
 <script lang="ts" setup>
-import UiSpinner from '@/components/ui/UiSpinner.vue'
 import { isHostOperationPending } from '@/libs/host'
 import { HOST_OPERATION } from '@/libs/xen-api/xen-api.enums'
 import type { XenApiHost } from '@/libs/xen-api/xen-api.types'
@@ -39,6 +38,7 @@ import VtsLayoutConsole from '@core/components/console/VtsLayoutConsole.vue'
 import VtsRemoteConsole from '@core/components/console/VtsRemoteConsole.vue'
 import VtsDivider from '@core/components/divider/VtsDivider.vue'
 import VtsStateHero from '@core/components/state-hero/VtsStateHero.vue'
+import UiLoader from '@core/components/ui/loader/UiLoader.vue'
 import { useUiStore } from '@core/stores/ui.store'
 import { computed, useTemplateRef } from 'vue'
 import { useI18n } from 'vue-i18n'
@@ -138,7 +138,7 @@ const sendCtrlAltDel = () => consoleElement.value?.sendCtrlAltDel()
   }
 }
 
-.spinner {
+.loader {
   color: var(--color-brand-txt-base);
   display: flex;
   margin: auto;