+++ @xen-orchestra/web-core/lib/components/drawer/VtsDrawerButton.vue
@@ -1,16 +1,17 @@
 <template>
-  <UiButton accent="brand" :busy="drawer?.isBusy.value" :variant size="medium">
+  <UiButton :accent :busy="drawer?.isBusy.value" :variant size="medium">
     <slot />
   </UiButton>
 </template>
 
 <script lang="ts" setup>
-import UiButton, { type ButtonVariant } from '@core/components/ui/button/UiButton.vue'
+import UiButton, { type ButtonAccent, type ButtonVariant } from '@core/components/ui/button/UiButton.vue'
 import { IK_DRAWER } from '@core/packages/drawer/types.ts'
 import { inject } from 'vue'
 
-defineProps<{
+const { accent = 'brand' } = defineProps<{
   variant: ButtonVariant
+  accent?: ButtonAccent
 }>()
 
 defineSlots<{