+++ @xen-orchestra/web/src/modules/pool/components/list/panel/PoolSidePanel.vue
@@ -4,12 +4,8 @@
       <VtsStateHero v-if="!arePoolsReady" format="panel" type="busy" size="medium" />
       <template v-else>
         <UiCard v-if="server.error === undefined" class="card-container">
-          <UiCardTitle>
-            {{ t('general-information') }}
-          </UiCardTitle>
+          <VtsCardObjectTitle :id="server.id" :label="server.label" icon="object:pool" />
           <div class="content">
-            <!-- ID -->
-            <VtsCodeSnippet :content="server.id" copy />
             <!-- Pool -->
             <VtsCardRowKeyValue>
               <template #key>{{ t('pool') }}</template>
@@ -167,7 +163,7 @@ import { useXoHostCollection } from '@/modules/host/remote-resources/use-xo-host
 import { useXoPoolCollection } from '@/modules/pool/remote-resources/use-xo-pool-collection.ts'
 import type { FrontXoServer } from '@/modules/server/remote-resources/use-xo-server-collection.ts'
 import VtsCardRowKeyValue from '@core/components/card/VtsCardRowKeyValue.vue'
-import VtsCodeSnippet from '@core/components/code-snippet/VtsCodeSnippet.vue'
+import VtsCardObjectTitle from '@core/components/card-object-title/VtsCardObjectTitle.vue'
 import VtsCopyButton from '@core/components/copy-button/VtsCopyButton.vue'
 import VtsIcon from '@core/components/icon/VtsIcon.vue'
 import VtsSidePanel from '@core/components/panel/VtsSidePanel.vue'